7.2 (#3)
* FIX Removing deprecated args from setup

* FIX Removing deprecated args from setup

* raise UserError in case requests response != 200
register ir.error
#178360

* try/except ConnectTimeout
#178409

* Replace config date_to to today in update_invoice_b2brouter_state()

* pyflakes

* Move b2b invoice fields to factura-e page invoice form

* Update the way that update the send invoice status. Not us interval of
days, download all invoices states that are in unfinished states.

Add a button to allow update invoices individually and not to way for
the cron process.

Add button to allow resend the invoice if it's needed. This will remove
the invoice send before and create a new one.

Add script to get the possible invoice states.

* Add the possibility to reset and forward invoice.
Reset, to remove, recreate and send the invoice.
Forward, to send the invoice again.

* Add new B2BRouter state.
Fix some bugs in  the sned, update B2BRouter API connection.

* Update view xpath structure with the change in account_invoice_facturae.

Task: #181070

---------

Co-authored-by: Raimon Esteve <raimonesteve@gmail.com>